fix a lot of imports in main app, organize legacy tests
