+++ tests/accounts/test_registration.py
