+++ tests/departments/api/v2/test_department_serializer.py
